requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
asn1crypto==1.0.0
async==0.6.2
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
capstone==4.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
checklib==0.2.8
Click==7.0
cryptography==2.9
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.23
intervaltree==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
multidict==4.6.1
numpy==1.17.2
packaging==19.2
pandas==0.25.1
parameterized==0.7.1
paramiko==2.6.0
pluggy==0.13.0
protobuf==3.11.3
psutil==5.6.3
psycopg2==2.8.3
psycopg2-binary==2.8.3
pwntools==2.2.1
py==1.8.0
pyasn1==0.4.7
pycparser==2.19
pycrypto==2.6.1
pyelftools==0.25
Pygments==2.4.2
pymongo==3.9.0
PyMySQL==0.9.3
PyMysqlDB==0.0.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pypandoc==1.4
pyparsing==2.4.2
PyPDF2==1.26.0
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
ROPGadget==5.8
rsa==4.0
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
selenium==3.141.0
six==1.14.0
sortedcontainers==1.5.10
swpag-client==0.3.7
toml==0.10.0
tox==3.14.0
typing-extensions==3.7.4.1
unicorn==1.0.1
urllib3==1.25.6
virtualenv==16.7.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
yaml-1.3==0.1.0
yarl==1.3.0
zipp==0.6.0
